requirements.txt
pytest==8.4.2
selene==2.0.0rc9
allure-pytest==2.15.0
allure-python-commons==2.15.0
selenium==4.35.0
python-dotenv==1.1.1
webdriver-manager==4.0.2
requests==2.32.5
telegram==0.0.1